% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca_utility.R
\name{calculate_outlyingness}
\alias{calculate_outlyingness}
\title{Calculate the outlyingness for the observations from one individual
Calculates the outlyingness matrix for one set of observations}
\usage{
calculate_outlyingness(
  x,
  t_true,
  grid,
  sigma,
  mean_estimator = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{x}{observation vector}

\item{t_true}{timestamps}

\item{grid}{the grid we're extrapolating to}

\item{sigma}{covariance matrix}
}
\description{
Calculate the outlyingness for the observations from one individual
Calculates the outlyingness matrix for one set of observations
}
