% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca.R
\name{cov.fun.hat2}
\alias{cov.fun.hat2}
\title{Estimates the covariance matrix}
\usage{
cov.fun.hat2(X, h, ma, ncov = 50, trace = FALSE)
}
\arguments{
\item{estimate_both_directions}{added by BB, do we really need to perform the smoothing step twice or
can we just assume the matrix is symmetric :)}
}
\description{
Estimates the covariance matrix
}
