% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{data_for_time_dynamic_fitting}
\alias{data_for_time_dynamic_fitting}
\title{STEP 3: MODELING THE SCORES
Fits the time dynamics (either robustly or non-robustly)
Transforms the data into the necessary structure}
\usage{
data_for_time_dynamic_fitting(design_df, estimated_scores)
}
\description{
STEP 3: MODELING THE SCORES
Fits the time dynamics (either robustly or non-robustly)
Transforms the data into the necessary structure
}
