% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca.R
\name{efpca}
\alias{efpca}
\title{Elliptical-FPCA main function}
\usage{
efpca(
  X,
  opt.h.cov,
  cutoff_outliers = 2.5,
  rho.param = 0.001,
  k = 3,
  s = 3,
  ncov = 50,
  max.kappa = 1000,
  prediction_grid = NULL,
  mean_estimator = function(x) mean(x, na.rm = TRUE)
)
}
\arguments{
\item{X}{number of values}
}
\description{
Elliptical-FPCA main function
}
