% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{estimate_mean}
\alias{estimate_mean}
\title{STEP 1: ESTIMATING THE MEAN FUNCTION}
\usage{
estimate_mean(modeling_data, method = "tensor-product", lambda_rqss = 20)
}
\description{
STEP 1: ESTIMATING THE MEAN FUNCTION
}
