% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{estimate_phi}
\alias{estimate_phi}
\title{STEP 2: ESTIMATING THE FPCs from the centered data}
\usage{
estimate_phi(
  modeling_df,
  centered_y,
  number_of_components = "pve",
  pve = 0.95,
  k = 2,
  cov_est = "MRCD"
)
}
\arguments{
\item{number_of_components}{algorithm to select the number of components, either
"kneedle" or "pve"}

\item{pve}{proportion of variance explained if \code{number_of_components = "pve"}, does not work if \code{robest = "MRCD"}}

\item{cov_est}{the robust estimator used to estimate the covariance (either \code{classic}, \code{PASS} or \code{MRCD}),
defaults to \code{classic}}

\item{y}{matrix of observations}
}
\value{
List of length two containing the eigenfunctions of the covariance function and the estimated mean
}
\description{
STEP 2: ESTIMATING THE FPCs from the centered data
}
