% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{estimate_scores}
\alias{estimate_scores}
\title{STEP 3: ESTIMATE THE SCORES
Estimate the scores using y and the eigenfunctions}
\usage{
estimate_scores(modeling_df, centered_y, phi, cov_center = NULL)
}
\arguments{
\item{phi}{the estimated eigenfunctions of the same length, i.e. dimensions \verb{ncol(y) x k}}

\item{y}{the matrix of observations}
}
\value{
A matrix of scores with \code{ncol(phi)} columns
}
\description{
STEP 3: ESTIMATE THE SCORES
Estimate the scores using y and the eigenfunctions
}
