% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca.R
\name{gtthat}
\alias{gtthat}
\title{This calculates gamma(t_0, t_0) as in equation (7) in the paper}
\usage{
gtthat(X, t0, h, b = 0.5, cc = 1.54764, initial.sc, max.it = 300, eps = 1e-10)
}
\description{
This calculates gamma(t_0, t_0) as in equation (7) in the paper
}
