% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca_utility.R
\name{k.epan}
\alias{k.epan}
\title{Epanechnikov kernel
Evaluates the Epanechnikov kernel for a given x}
\usage{
k.epan(x)
}
\arguments{
\item{x}{the value}
}
\description{
Epanechnikov kernel
Evaluates the Epanechnikov kernel for a given x
}
