% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mean_abs_dev}
\alias{mean_abs_dev}
\title{Calculate the mean absolute deviation}
\usage{
mean_abs_dev(a, b)
}
\arguments{
\item{a}{true values}

\item{b}{estimated values of same length as \code{a}}
}
\value{
The mean absolute deviation between \code{a} and \code{b}
}
\description{
Calculate the mean absolute deviation
}
