% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality}
\alias{mortality}
\title{Mortality data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 207518 rows and 8 columns.
}
\usage{
mortality
}
\description{
Mortality data
}
\keyword{datasets}
