% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_eigenanalysis.R
\name{performPASSeigenAnalysis}
\alias{performPASSeigenAnalysis}
\title{Taken from https://github.com/kwgit/PASSrobustFPCA
Calculate the eigenvalues for the PASS covariance matrix}
\usage{
performPASSeigenAnalysis(
  fdata,
  data.cov = "data",
  trim = 0,
  tol = 1e-10,
  max.itr = 1e+20
)
}
\description{
Taken from https://github.com/kwgit/PASSrobustFPCA
Calculate the eigenvalues for the PASS covariance matrix
}
