% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plots a matrix using geom_tile()}
\usage{
plot_matrix(cc, index = NULL)
}
\arguments{
\item{cc}{the matrix that should be plotted}

\item{index}{the index that should be used for the x- and y-axis of the plot}
}
\value{
A ggplot
}
\description{
Plots a matrix using geom_tile()
}
