% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{predict_grid}
\alias{predict_grid}
\title{PREDICT TRAJECTORIES FOR THE WHOLE GRID
Predict the whole grid of information from the fitted model}
\usage{
predict_grid(
  ss,
  tt,
  subject_ids,
  models,
  design_df = NULL,
  y = NULL,
  phi,
  mean_fn_model,
  cov_center = NULL
)
}
\arguments{
\item{ss}{Frequency domain indices}

\item{tt}{Complete time domain indices}

\item{subject_ids}{Which subjects to predict}

\item{models}{the models for the time dynamics (as returned by \code{fit_time_dynamics()})}

\item{phi}{the estimated eigenfunctions}

\item{mean_fn_model}{the model for the estimated mean function}
}
\value{
The predicted y-values for all combinations of ss and tt and subject_id
in a data frame.
}
\description{
PREDICT TRAJECTORIES FOR THE WHOLE GRID
Predict the whole grid of information from the fitted model
}
