% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{predict_grid_nonrobust}
\alias{predict_grid_nonrobust}
\title{Non-robust prediction}
\usage{
predict_grid_nonrobust(ss, tt, subject_ids, models, phi, mean_fn_model)
}
\description{
Non-robust prediction
}
