% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efpca.R
\name{predict_score_functions}
\alias{predict_score_functions}
\title{Predict smoothed trajectories from the estimated covariance model}
\usage{
predict_score_functions(
  X,
  cov.fun,
  tt,
  flagged_observations,
  s = 20,
  rho = 0,
  pve = 0.9
)
}
\arguments{
\item{X}{list of lists of length two with measurements and timepoints}

\item{cov.fun}{smoothed covariance matrix (length(tt) x length(tt) matrix)}

\item{tt}{timepoints to predict on}

\item{s}{maximum number of components}

\item{rho}{number of measurements}

\item{pve}{pve to decide on \code{k}, the number of components}
}
\description{
Predict smoothed trajectories from the estimated covariance model
}
