% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rmse}
\alias{rmse}
\title{Calculate the RMSE}
\usage{
rmse(a, b)
}
\arguments{
\item{a}{true values}

\item{b}{estimated values of same length as \code{a}}
}
\value{
The RMSE between \code{a} and \code{b}
}
\description{
Calculate the RMSE
}
