% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uvvis}
\alias{uvvis}
\title{Uvvis Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18795 rows and 11 columns.
}
\usage{
uvvis
}
\description{
Uvvis Data
}
\keyword{datasets}
